% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpc.R
\name{callGRPCMethod}
\alias{callGRPCMethod}
\title{Issues a gRPC call to the Google Genomics service and returns the
response.}
\usage{
callGRPCMethod(methodName, request)
}
\arguments{
\item{methodName}{The RPC method name.}

\item{request}{The request object for the RPC, either as a JSON object
generated from \code{\link[rjson]{rjson}}, or as a
\code{\link[RProtoBuf]{RProtoBuf}} object modified from the
default instance obtained from
\code{\link{getRProtoBufDefaultObject}}.}
}
\value{
The raw response converted from JSON to an R object, or the
        RProtoBuf object if the request was an RProtoBuf object.
}
\description{
Needs gRPC support at package build time and the RProtoBuf package.
See package README for instructions on installing gRPC.
}
\details{
In general, use higher level methods such as getReads and getVariants
instead.
}
\examples{
# Authenticated on package load from the env variable GOOGLE_API_KEY.
if (isGRPCAvailable()) {
  request <- list(readGroupSetIds=list("CMvnhpKTFhDnk4_9zcKO3_YB"),
                  referenceName="22",
                  start=16051400, end=16051500, pageToken=NULL)
  reads <- callGRPCMethod("SearchReads", request)
  summary(reads)
} else {
  message("gRPC support is disabled; package was not compiled with GRPC")
}
}
\seealso{
Other page fetch functions: \code{\link{getReadsPage}},
  \code{\link{getSearchPage}},
  \code{\link{getVariantsPage}}
}
