% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReadData}
\alias{getReadData}
\title{Get reads data from Google Genomics.}
\usage{
getReadData(chromosome = "chr19", start = 45411941, end = 45412079,
  readsetId = "CJ_ppJ-WCxDxrtDr5fGIhBA=",
  endpoint = "https://www.googleapis.com/genomics/v1beta/",
  pageToken = NULL)
}
\arguments{
\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome.}

\item{end}{End position on the chromosome.}

\item{readsetId}{The readset ID.}

\item{endpoint}{The Google API endpoint. You should not be changing this.}

\item{pageToken}{The page token. This can be NULL (default) for the first page.}
}
\description{
By default, this function encompasses 2 chromosome positions which relate to ApoE
(http://www.snpedia.com/index.php/Rs429358 and http://www.snpedia.com/index.php/Rs7412)
}

