% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{getReadsPage}
\alias{getReadsPage}
\title{Get one page of reads from Google Genomics.}
\usage{
getReadsPage(readGroupSetId = "CMvnhpKTFhDnk4_9zcKO3_YB", chromosome = "22",
  start = 16051400, end = 16051500, fields = NULL, pageToken = NULL)
}
\arguments{
\item{readGroupSetId}{The read group set ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{pageToken}{The page token. This can be NULL (default) for the first
page.}
}
\value{
A two-element list is returned by the function.

    reads: A list of R objects corresponding to the JSON objects returned
              by the Google Genomics Reads API.

    nextPageToken: The token to be used to retrieve the next page of
                   results, if applicable.
}
\description{
In general, use the getReads method instead.  It calls this method,
returning reads from all of the pages that comprise the requested
genomic range.
}
\details{
By default, this function gets reads for a small genomic region for one
sample in 1,000 Genomes.

Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}
\examples{
# Authenticated on package load from the env variable GOOGLE_API_KEY.
readsPage <- getReadsPage()
summary(readsPage)
summary(readsPage$reads[[1]])
}
\seealso{
Other page fetch functions: \code{\link{callGRPCMethod}},
  \code{\link{getSearchPage}},
  \code{\link{getVariantsPage}}
}
