% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpc.R
\name{getRProtoBufDefaultObject}
\alias{getRProtoBufDefaultObject}
\title{Returns a protocol buffer Message object from
\code{\link[RProtoBuf]{RProtoBuf}}.}
\usage{
getRProtoBufDefaultObject(fullyQualifiedName)
}
\arguments{
\item{fullyQualifiedName}{Type of the message object to return.}
}
\value{
Default instance of the Message.
}
\description{
Needs gRPC support at package build time and the RProtoBuf package.
See package README for instructions on installing gRPC.
}
\examples{
if (isGRPCAvailable()) {
  getRProtoBufDefaultObject("google.genomics.v1.SearchReadsRequest")
}
}
