% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{getSearchPage}
\alias{getSearchPage}
\title{Get one page of search results for a particular entity type from Google
  Genomics.}
\usage{
getSearchPage(entityType, body, fields, pageToken)
}
\arguments{
\item{entityType}{Entities with a search API such as reads, variants,
variantSets, etc...}

\item{body}{The body of the message to POST to the search endpoint.}

\item{fields}{The fields to be returned in the search response.}

\item{pageToken}{The page token. This can be NULL for the first page.}
}
\value{
The raw response converted from JSON to an R object.
}
\description{
In general, use higher level methods such as getReads and getVariants
instead.
}
\examples{
# Authenticated on package load from the env variable GOOGLE_API_KEY.
body <- list(readGroupSetIds=list("CMvnhpKTFhDnk4_9zcKO3_YB"),
             referenceName="22",
             start=16051400, end=16051500, pageToken=NULL)
reads <- getSearchPage("reads", body, NULL, NULL)
summary(reads)
}
\seealso{
Other page fetch functions: \code{\link{getReadsPage}};
  \code{\link{getVariantsPage}}
}

