% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVariantData}
\alias{getVariantData}
\title{Get variants data from Google Genomics.}
\usage{
getVariantData(datasetId = "10473108253681171589", chromosome = "22",
  start = 16051400, end = 16051500,
  endpoint = "https://www.googleapis.com/genomics/v1beta/",
  pageToken = NULL)
}
\arguments{
\item{datasetId}{The dataset ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome.}

\item{end}{End position on the chromosome.}

\item{endpoint}{The Google API endpoint. You should not be changing this.}

\item{pageToken}{The page token. This can be NULL (default) for the first page.}
}
\description{
By default, this function gets variant data from a small section of 1000 genomes
}

