% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{getVariantsPage}
\alias{getVariantsPage}
\title{Get one page of variants from Google Genomics.}
\usage{
getVariantsPage(datasetId = "10473108253681171589", chromosome = "22",
  start = 16051400, end = 16051500, fields = NULL, pageToken = NULL)
}
\arguments{
\item{datasetId}{The dataset ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{pageToken}{The page token. This can be NULL (default) for the first
page.}
}
\value{
A two-element list is returned by the function.

    variants: A list of R objects corresponding to the JSON objects returned
              by the Google Genomics Variants API.

    nextPageToken: The token to be used to retrieve the next page of
                   results, if applicable.
}
\description{
In general, use the getVariants method instead.  It calls this method,
returning variants from all of the pages that comprise the requested
genomic range.
}
\details{
By default, this function gets variants from a small section of 1000
Genomes phase 1 variants.
}
\examples{
# Authenticated on package load from the env variable GOOGLE_API_KEY.
variantsPage <- getVariantsPage()
summary(variantsPage)
summary(variantsPage$variants[[1]])
}
\seealso{
Other page fetch functions: \code{\link{callGRPCMethod}},
  \code{\link{getReadsPage}}, \code{\link{getSearchPage}}
}
