% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpc.R
\name{isGRPCAvailable}
\alias{isGRPCAvailable}
\title{Returns if this package was built with gRPC support.}
\usage{
isGRPCAvailable()
}
\value{
TRUE iff the package was built with gRPC support.
}
\description{
Returns if this package was built with gRPC support.
}
