% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reads.R
\name{readsToGAlignments}
\alias{readsToGAlignments}
\title{Convert reads to GAlignments.}
\usage{
readsToGAlignments(reads, oneBasedCoord = TRUE, slStyle = "UCSC")
}
\arguments{
\item{reads}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Reads API.}

\item{oneBasedCoord}{Convert genomic positions to 1-based coordinates.}

\item{slStyle}{The style for seqnames (chrN or N or...).  Default is UCSC.}
}
\value{
\link[GenomicAlignments]{GAlignments}
}
\description{
Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}
\examples{
# Authenticated on package load from the env variable GOOGLE_API_KEY.
alignments1 <- getReads(converter=readsToGAlignments)
summary(alignments1)
alignments2 <- readsToGAlignments(getReads())
print(identical(alignments1, alignments2))
}

