% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{readStore}
\alias{readStore}
\title{Read Store.}
\format{An R environment.}
\usage{
readStore
}
\description{
Environment where the most recently read alignments and reads are stored.
  Reads can be accessed by \code{readStore$reads}.
  Alignments can be accessed by \code{readStore$alignments}.
}
\keyword{datasets}

