% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{variantStore}
\alias{variantStore}
\title{Variants Store.}
\format{\preformatted{<environment: 0x7fd36bf1bd08> 
}}
\usage{
variantStore
}
\description{
Environment where the most recently read variants are stored.
  Variants can be accessed by \code{variantStore$variants}.
  Detailed variants data can be accessed by \code{variantStore$variantsData}.
}
\keyword{datasets}

