% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variants.R
\name{variantsToVCF}
\alias{variantsToVCF}
\title{Convert variants to VCF.}
\usage{
variantsToVCF(variants)
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}
}
\value{
\link[VariantAnnotation]{VCF}
}
\description{
Convert variants to VCF.
}
\seealso{
Other variants converter functions: \code{\link{variantsToGRanges}};
  \code{\link{variantsToVRanges}}
}

