% Generated by roxygen2 (4.0.2): do not edit by hand
\name{variantsToVRanges}
\alias{variantsToVRanges}
\title{Convert variants to VRanges.}
\usage{
variantsToVRanges(variants)
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}
}
\value{
\link[VariantAnnotation]{VRanges}
}
\description{
Note that genomic coordinates are converted from 0-based to 1-based.
}

