% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authenticate}
\alias{authenticate}
\title{Configure how to authenticate for Google Genomics API.}
\usage{
authenticate(file, clientId, clientSecret, invokeBrowser = TRUE, apiKey)
}
\arguments{
\item{file}{Client secrets file obtained from Google Developer Console.
If this file is not present, clientId and clientSecret must be provided.}

\item{clientId}{Client ID from Google Developer Console.}

\item{clientSecret}{Client Secret from Google Developer Console.}

\item{invokeBrowser}{If true, the default browser is invoked with the auth URL (default).
If false, a URL is output which needs to be copy pasted in a browser.
With both the options, you will still need to login to your Google account if not logged
in already, and paste back the token key.}

\item{apiKey}{Public API key that can be used to call the Genomics API for public datasets.
Providing this key overrides all other arguments.}
}
\description{
Follow the sign up instructions at \url{https://developers.google.com/genomics} to download
  the client secrets file, or note the clientId and clientSecret pair.
}

