% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Configure how to authenticate for Google Genomics API.}
\usage{
authenticate(file, clientId, clientSecret, invokeBrowser, apiKey)
}
\arguments{
\item{file}{Client secrets file obtained from Google Developer Console. This
file could be for a native application or a service account. If this file
is not present, clientId and clientSecret must be provided for native
application credentials. Service account support needs version 0.1-2 or
greater of \href{https://github.com/s-u/PKI}{PKI}.}

\item{clientId}{Client ID from Google Developer Console, overridden if file
is provided.}

\item{clientSecret}{Client Secret from Google Developer Console, overridden
if file is provided.}

\item{invokeBrowser}{If TRUE or not provided, the default browser is invoked
with the auth URL iff the \code{\link[httpuv]{httpuv}} package is
installed (suggested). If FALSE, a URL is output which needs to be copy
pasted in a browser, and the resulting token needs to be pasted back into
the R session. With both the options, you will still need to login to your
Google account if not logged in already.}

\item{apiKey}{Public API key that can be used to call the Genomics API for
public datasets. This method of authentication does not need you to login
to your Google account. Providing this key overrides all other arguments.}
}
\value{
NULL (silently) if successful.
}
\description{
Follow the sign up instructions at \url{https://cloud.google.com/genomics/install-genomics-tools#authenticate}
  to download the client secrets file, or note the clientId and clientSecret
  pair.
}
\examples{
apiKey <- Sys.getenv("GOOGLE_API_KEY")
if (!is.na(apiKey) && nchar(apiKey)>0) {
  authenticate(apiKey=apiKey)
}
\dontrun{
authenticate(file="clientSecrets.json")
authenticate(file="clientSecrets.json", invokeBrowser=FALSE)
authenticate(clientId="abc", clientSecret="xyz", invokeBrowser=FALSE)
}
}

