% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSearchPage}
\alias{getSearchPage}
\title{Get one page of search results for a particular entity type from Google Genomics.}
\usage{
getSearchPage(entityType, body, fields, pageToken)
}
\arguments{
\item{entityType}{Entities with a search API such as reads, variants, variantSets, etc...}

\item{body}{The body of the message to POST to the search endpoint.}

\item{fields}{The fields to be returned in the search response.}

\item{pageToken}{The page token. This can be NULL for the first page.}
}
\value{
The raw response converted from JSON to an R object.
}
\description{
In general, use higher level methods such as getReads and getVariants
instead.
}

