% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{getVariantCalls}
\alias{getVariantCalls}
\title{Elaborate the result of getVariants as a VRanges with all
calls for all samples}
\usage{
getVariantCalls(datasetId = "10473108253681171589", chromosome = "22",
  start = 16051400, end = 16051500, fields = NULL, converter = c,
  oneBasedCoord = TRUE, nullaction = "stop")
}
\arguments{
\item{datasetId}{The dataset ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{converter}{A function that takes a list of variant R objects and
returns them converted to the desired type.}

\item{oneBasedCoord}{Convert returned addresses to 1-based address system}

\item{nullaction}{either \code{"stop"} or \code{"warn"} telling how to deal with event in which request yields no variants; for \code{"warn"} we return NULL}
}
\description{
Elaborate the result of getVariants as a VRanges with all
calls for all samples
}
\examples{
# default to generate VRanges
getVariantCalls()
}

