% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVariants}
\alias{getVariants}
\title{Get variants from Google Genomics.}
\usage{
getVariants(datasetId = "10473108253681171589", chromosome = "22",
  start = 16051400, end = 16051500, oneBasedCoord = FALSE,
  fields = NULL, converter = c)
}
\arguments{
\item{datasetId}{The dataset ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome.}

\item{end}{End position on the chromosome.}

\item{oneBasedCoord}{Interpret start and end as 1-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{converter}{A function that takes a list of variant R objects and returns
them converted to the desired type.}
}
\value{
By default, the return value is a list of R objects corresponding to the JSON objects
 returned by the Google Genomics Variants API.  If a converter is passed,
 object(s) of the type returned by the converter will be returned by this function.
}
\description{
This function will return all of the variants that comprise the requested
genomic range, iterating over paginated results as necessary.
}
\details{
By default, this function gets variants from a small section of 1000
Genomes phase 1 variants.

Optionally pass a converter as appropriate for your use case.  By passing it
to this method, only the converted objects will be accumulated in memory. The
converter function should return an empty container of the desired type
if called without any arguments.

Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}

