% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVariantsPage}
\alias{getVariantsPage}
\title{Get one page of variants from Google Genomics.}
\usage{
getVariantsPage(datasetId = "10473108253681171589", chromosome = "22",
  start = 16051400, end = 16051500, oneBasedCoord = FALSE,
  fields = NULL, pageToken = NULL)
}
\arguments{
\item{datasetId}{The dataset ID.}

\item{chromosome}{The chromosome.}

\item{start}{Start position on the chromosome.}

\item{end}{End position on the chromosome.}

\item{oneBasedCoord}{Interpret start and end as 1-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{pageToken}{The page token. This can be NULL (default) for the first page.}
}
\value{
a two-element list
    variants: a list of R objects corresponding to the JSON objects returned
              by the Google Genomics Variants API
    nextPageToken: the token to be used to retrieve the next page of
                   results, if applicable
}
\description{
In general, use the getVariants method instead.  It calls this method,
returning variants from all of the pages that comprise the requested
genomic range.
}
\details{
By default, this function gets variants from a small section of 1000
Genomes phase 1 variants.
}

