% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readsToGAlignments}
\alias{readsToGAlignments}
\title{Convert reads to GAlignments.}
\usage{
readsToGAlignments(reads)
}
\arguments{
\item{reads}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}
}
\value{
\link[GAlignments]{GAlignments}
}
\description{
Convert reads to GAlignments.
}

