% Generated by roxygen2 (4.0.2): do not edit by hand
\name{variantsToGRanges}
\alias{variantsToGRanges}
\title{Convert variants to GRanges.}
\usage{
variantsToGRanges(variants, oneBasedCoord = TRUE, slStyle = "UCSC")
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}

\item{oneBasedCoord}{Convert genomic positions to 1-based coordinates.}

\item{slStyle}{The style for seqnames (chrN or N or...).  Default is UCSC.}
}
\value{
\link[GenomicRanges]{GRanges}
}
\description{
Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}

