% Generated by roxygen2 (4.0.2): do not edit by hand
\name{variantsToGRanges}
\alias{variantsToGRanges}
\title{Convert variants to GRanges.}
\usage{
variantsToGRanges(variants)
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}
}
\value{
GRanges
}
\description{
Note that genomic coordinates are converted from 0-based to 1-based.
}
\references{
GRanges
}

