% Generated by roxygen2 (4.0.2): do not edit by hand
\name{variantsToVCF}
\alias{variantsToVCF}
\title{Convert variants to VCF.}
\usage{
variantsToVCF(variants)
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the Google Genomics Variants API.}
}
\value{
\link[VariantAnnotation]{VCF}
}
\description{
Convert variants to VCF.
}

